Refactor test suite
