+++ .gitignore
@@ -1,7 +1,4 @@
-*~
-Tests/app/cache/
-composer.lock
-phpunit.xml
-vendor/
-Tests/web
-var/
+/composer.lock
+/phpunit.xml
+/vendor/
+/var/
