+++ Tests/Integration/config/packages/framework.yaml
@@ -1,0 +1,14 @@
+framework:
+    secret: '%env(APP_SECRET)%'
+    #csrf_protection: true
+    #http_method_override: true
+
+    # Enables session support. Note that the session will ONLY be started if you read or write from it.
+    # Remove or comment this section to explicitly disable session support.
+    session:
+        handler_id: null
+
+    #esi: true
+    #fragments: true
+    php_errors:
+        log: true
