+++ Tests/Integration/config/packages/presta_sitemap.yaml
@@ -1,0 +1,3 @@
+presta_sitemap:
+    default_section: static
+    dump_directory: "%kernel.project_dir%/public"
