+++ Tests/Integration/config/routes/presta_sitemap.yaml
@@ -1,0 +1,2 @@
+presta_sitemap:
+    resource: "@PrestaSitemapBundle/Resources/config/routing.yml"
