+++ Tests/Integration/config/routes/xml.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+
+<routes xmlns="http://symfony.com/schema/routing"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://symfony.com/schema/routing http://symfony.com/schema/routing/routing-1.0.xsd">
+
+    <route id="xml" path="/xml">
+        <default key="_controller">Presta\SitemapBundle\Tests\Integration\Controller\FormatController::xml</default>
+        <option key="sitemap">
+            {"priority":"0.7", "changefreq":"weekly", "section":"static"}
+        </option>
+    </route>
+
+</routes>
