--- Tests/app/config.sf3.yml
@@ -1,21 +1,0 @@
-framework:
-    secret:          secret
-    test: ~
-    router:          { resource: "%kernel.root_dir%/routing.yml" }
-    form:            true
-    csrf_protection: true
-    validation:      { enable_annotations: true }
-    session:
-        storage_id: session.storage.filesystem
-
-security:
-    providers:
-        in_memory:
-            memory: ~
-    firewalls:
-        dev:
-            pattern: ^/(_(profiler|wdt)|css|images|js)/
-            security: false
-        main:
-            anonymous: ~
-            logout_on_user_change: true
