--- Tests/app/config.yml
@@ -1,20 +1,0 @@
-framework:
-    secret:          secret
-    test: ~
-    router:          { resource: "%kernel.project_dir%/Tests/app/routing.yml" }
-    form:            true
-    csrf_protection: true
-    validation:      { enable_annotations: true }
-    session:
-        storage_id: session.storage.filesystem
-
-security:
-    providers:
-        in_memory:
-            memory: ~
-    firewalls:
-        dev:
-            pattern: ^/(_(profiler|wdt)|css|images|js)/
-            security: false
-        main:
-            anonymous: ~
