--- Tests/fixtures/sitemap.video.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
-        xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-  <url>
-    <loc>http://sitemap.php54.local/page_video1/</loc>
-    <lastmod>2020-03-01T17:48:38+01:00</lastmod>
-    <changefreq>daily</changefreq>
-    <priority>1.0</priority>
-    <video:video>
-      <video:thumbnail_loc>http://sitemap.php54.local/page_video1/thumbnail_loc?a=b&amp;b=c</video:thumbnail_loc>
-      <video:title><![CDATA[Title & spécial chars]]></video:title>
-      <video:description><![CDATA[The description & spécial chars]]></video:description>
-      <video:content_loc>http://sitemap.php54.local/page_video1/content?format=mov&amp;a=b</video:content_loc>
-      <video:gallery_loc title="Gallery title &amp; spécial chars">http://sitemap.php54.local/page_video1/gallery_loc/?p=1&amp;sort=desc</video:gallery_loc>
-    </video:video>
-  </url>
-</urlset>
