--- Tests/fixtures/sitemap.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<sitemapindex xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-              xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9 http://www.sitemaps.org/schemas/sitemap/0.9/siteindex.xsd"
-              xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-    <sitemap>
-        <loc>http://sitemap.php54.local/sitemap.audio.xml</loc>
-        <lastmod>2013-08-04T15:10:01+03:00</lastmod>
-    </sitemap>
-    <sitemap>
-        <loc>http://sitemap.php54.local/sitemap.video.xml</loc>
-        <lastmod>2013-08-04T15:05:01+03:00</lastmod>
-    </sitemap>
-</sitemapindex>
